requirements.txt
Django>=4.2,<5.0
djangorestframework
django-jazzmin
Pillow
psycopg2-binary>=2.9.9
django-filter>=23.5
djangorestframework-simplejwt
drf-spectacular
djangorestframework